separate services directory, move DocumentService there
